update: canister version numbers and add health check functionality
This commit updates the version number of both FileScalingManager and FileStorage canisters from 1 to 2. It also adds a new `Health` type to the `types.mo` module, which is used in the `CanisterInfo` type. The `FileScalingManager` now has a new private function called `update_health()` that updates the health status of all child canisters every 10 minutes using the newly added `get_health()` query method in the `FileStorage` actor. Finally, this commit includes changes to some test files that reflect these updates.